requirement hotfix

+++ src/services/cart.service.js
@@ -30,18 +30,24 @@ class CartService {
   finalizePurchase = async(cid, purchaser)=>{
     let cart = new CartDTO(this.findcartById(cid))
     let totalAmount = 0;
+    let productsWithoutStock = [];
+
     cart.products.forEach(async (item, idx, pArray) =>{ 
       const product = new ProductDTO(await productService.findproductById(item.product))
       if(item.quantity > product.stock)
-        pArray.splice(idx,1)
+        productsWithoutStock = [...productsWithoutStock, pArray.splice(idx,1)[0].product]
       else{
         product.stock =- item.quantity
         await productService.updateProduct(product._id, product)
         totalAmount += (product.price * item.quantity)
       }
     })
-    await ticketService.createTicket(totalAmount, purchaser)
-    await cartRepository.delete(cid)
+    if(productsWithoutStock.length > 0)
+      return productsWithoutStock
+    else{
+      await ticketService.createTicket(totalAmount, purchaser)
+      await cartRepository.delete(cid)
+    }
   }
 }
